Mark reserved-for-write tvars with a flag in packed pointer, unblock reads
To correctly commit the result of a write transaction, we need to ensure that
the input tvar values do not change until the write transaction finishes its
commit. If a particular write transaction has its inputs overwritten before it enters the commit phase, it can simply re-check its inputs, detect the changes, and roll back. During the actual commit, however, a write transaction needs to ensure that another transaction doesn't change its inputs as it is exposing its new values to the world.

In the past, we reserved a special magical pointer version that did not
actually point to any interesting data. Instead, it only signified that the
tvar was undergoing an update. This successfully communicated to other write
transactions that the tvar was updating, allowing them to spin until the
commit was finished. Because the original pointer version was swapped out,
however, this also imposed this blocking on read transactions.

This new scheme instead forces all version pointees to be aligned on at least
a two-byte boundary. This means that the low bit of any pointer pointing to a
version pointee will be 0. This allows us to use this low bit to indicate
whether the pointed-to version is reserved for writes or not. Transactions
which are not yet writing may now ignore this low bit and continue, while
transactions that are attempting to commit can be asked to wait until the
reserving transaction is finished.