Clean up txn timestamps and reduce number of dereferences
Previously, we had all version objects commited within a write transaction
point to the same timestamp object, which we would update with a release write
at the end of the transaction. That is unnecessary for the properties we wish
to provide, and the single write to update the timestamps is not worth the
indirection on every read. This change embeds the timestamp in the version
objects and causes write transactions to update every committed object with
that transaction number.