Extend tvars to allow flexible array members
Before, all TVars had to be of fixed size. Now, all TVars are assumed to be
flexible arrays, with some fixed-size header and a resizable flexible array
afterwards, the size of which is provided by the fixed size header. TVars that
do not need this flexible array are represented as being 100% header with a
flexible array size of 0. While not the common case, flexible arrays are
essential for many common data structures (for instance, the backing storage
of a hash table, vector, or skip list).