Overhaul captured tvar borrowing to allow for more static checking
Previously, the Rust reference invariants of CapturedTVarRef and
CapturedTVarMut were handled entirely dynamically. This was annoying because
problems could show up only at runtime. This change forces only one
CapturedTVar to exist at once and causes CapturedTVarRef and CapturedTVarMut
to borrow the CapturedTVar in the appropriate fashion. This allows the borrow
checker to statically enforce the borrow invariants. The stricter dynamic
constraints on the use of CapturedTVars is a bit of a bummer; one could
imagine providing mutable and immutable variants for CapturedTVars and
providing enforcement parallel to references for checking. We will think about
that for the future.