Make initial versions of tvars reserve a version in the txn write time
Before this change, tvar versions created for VersionedTVar::new would be
marked with a version of 0. This would be unambiguously fine if this function
were called only before main. However, due to dynamic allocation of tvars and
lazy_static, the initial version of a tvar may be placed within the tvar at
transaction times other than 0. To reflect this reality, stamp the initial
version of the tvar with the current write time.