Allow TVars to be initially unfilled, preventing allocation before usage
This change allows TVars to have an "unfilled" state, in which their version
pointer points to null. This prevents from having to perform an allocation
before the TVar's first use. This allows TVars to be created without use of
lazy_static and allows them to take up only the space for the TVar itself
before the TVar is filled, possibly lazily right before its first use.