Get rid of the extraneous TVarVersionClone
The original idea of TVarVersionClone was to allow tvars to be used with types
that did not want to be clonable in general, but did want to be clonable in
the specific case of tvar version duplication. I increasingly think that this
was a mistake; we want the clone for tvar versions to just look like a regular
clone, and any type that shouldn't be cloned by a user probably shouldn't be
cloned by tvar version creation.