Trailing array inits are now slice refs, not Vecs
Initially, I made the initializers for trailing arrays on VersionedTVars
vectors of the member item. This was done so that the constructor could take
ownership of the list of items to initialize the new TVar with, sidestepping
requiring a Clone on the array member type. Now that I've decided that
TVarClone isn't worth the trouble, however, requiring a Vec just requires an
additional allocation for no reason. This change makes the initializer a
non-mutable slice ref, which allows callers to just pass in a slice to an
immutable array in their stack frame or in a const global.